+++ doc/user/Changelog.md
@@ -113,6 +113,11 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Improve how vehicle rental networks and vehicles are presented in the debug client [#7888](https://github.com/opentripplanner/OpenTripPlanner/pull/7888)
 - Direction of line restriction in SIRI-SX [#7903](https://github.com/opentripplanner/OpenTripPlanner/pull/7903)
 - Support for SIRI partial replacedBy [#7873](https://github.com/opentripplanner/OpenTripPlanner/pull/7873)
+- Expose scheduled quay [#7739](https://github.com/opentripplanner/OpenTripPlanner/pull/7739)
+- Document the transaction framework and cleanup module test [#7894](https://github.com/opentripplanner/OpenTripPlanner/pull/7894)
+- Frequency-based vehicle position matching [#7927](https://github.com/opentripplanner/OpenTripPlanner/pull/7927)
+- Move per-network GBFS configuration into a shared gbfs section of otp-config.json [#7886](https://github.com/opentripplanner/OpenTripPlanner/pull/7886)
+- Compact node storage to reduce graph build memory [#7925](https://github.com/opentripplanner/OpenTripPlanner/pull/7925)
 [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)
 
 ## 2.9.0 (2026-03-18)